+++ src/pages/Mypage/Mypage.jsx
@@ -6,21 +6,9 @@ import { faMagnifyingGlass, faBars } from "@fortawesome/free-solid-svg-icons";
 import Header from "../../component/Header";
 import { HeartOutlined, HeartFilled } from "@ant-design/icons";
 import { render } from "@testing-library/react";
+import MypageItem from "../../component/MypageItem";
 
 function MyPage() {
-  const [Like, setLike] = useState(false);
-  const [cnt, setCnt] = useState(0);
-
-  const handleLike = () => {
-    if (Like === true) {
-      setLike(false);
-      setCnt((prev) => prev - 1);
-    } else {
-      setLike(true);
-      setCnt((prev) => prev + 1);
-    }
-  };
-
   return (
     <>
       <Header menu="myPage" />
@@ -47,105 +35,12 @@ function MyPage() {
             </div>
           </div>
           <S.writeList>
-            <div className="settingWriting">
-              <div className="settingWritingConent">
-                <div className="imgSetting">
-                  <S.writeImg src="https://t1.daumcdn.net/cfile/tistory/186E94454F7E700720" />
-                </div>
-                <S.writingContent>
-                  <span>김새미 인성관련 이슈</span>
-                  <br />
-                  <span className="content">
-                    마라탕집에서 꿔바로우 훔쳐 도망.
-                  </span>
-                  <div className="DateRecommend">
-                    <span className="date">1시간 전</span>
-                    <div>
-                      {Like === true ? (
-                        <HeartFilled
-                          className="heart"
-                          onClick={handleLike}
-                          style={{ color: "red", fontSize: "15px" }}
-                        />
-                      ) : (
-                        <HeartOutlined
-                          className="heart"
-                          style={{ color: "#a1a1a1" }}
-                          onClick={handleLike}
-                        />
-                      )}
-                      <span className="counter">{cnt}</span>
-                    </div>
-                  </div>
-                </S.writingContent>
-              </div>
-            </div>
-            <div className="settingWriting">
-              <div className="settingWritingConent">
-                <div className="imgSetting">
-                  <S.writeImg src="https://t1.daumcdn.net/cfile/tistory/186E94454F7E700720" />
-                </div>
-                <S.writingContent>
-                  <span>김새미 인성관련 이슈</span>
-                  <br />
-                  <span className="content">
-                    마라탕집에서 꿔바로우 훔쳐 도망.
-                  </span>
-                  <div className="DateRecommend">
-                    <span className="date">1시간 전</span>
-                    <div>
-                      {Like === true ? (
-                        <HeartFilled
-                          className="heart"
-                          onClick={handleLike}
-                          style={{ color: "red", fontSize: "15px" }}
-                        />
-                      ) : (
-                        <HeartOutlined
-                          className="heart"
-                          style={{ color: "#a1a1a1" }}
-                          onClick={handleLike}
-                        />
-                      )}
-                      <span className="counter">{cnt}</span>
-                    </div>
-                  </div>
-                </S.writingContent>
-              </div>
-            </div>
-            <div className="settingWriting">
-              <div className="settingWritingConent">
-                <div className="imgSetting">
-                  <S.writeImg src="https://t1.daumcdn.net/cfile/tistory/186E94454F7E700720" />
-                </div>
-                <S.writingContent>
-                  <span>김새미 인성관련 이슈</span>
-                  <br />
-                  <span className="content">
-                    마라탕집에서 꿔바로우 훔쳐 도망.
-                  </span>
-                  <div className="DateRecommend">
-                    <span className="date">1시간 전</span>
-                    <div>
-                      {Like === true ? (
-                        <HeartFilled
-                          className="heart"
-                          onClick={handleLike}
-                          style={{ color: "red", fontSize: "15px" }}
-                        />
-                      ) : (
-                        <HeartOutlined
-                          className="heart"
-                          style={{ color: "#a1a1a1" }}
-                          onClick={handleLike}
-                        />
-                      )}
-                      <span className="counter">{cnt}</span>
-                    </div>
-                  </div>
-                </S.writingContent>
-              </div>
-            </div>
+            <MypageItem />
+            <MypageItem />
+            <MypageItem />
+            <MypageItem />
+            <MypageItem />
+            
           </S.writeList>
         </S.ArticlesList>
       </S.ContentCenter>
